requirements.txt
amqp==5.2.0
anyio==3.7.1
asgiref==3.6.0
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.1.2
Automat==0.8.0
backports.zoneinfo==0.2.1
billiard==4.2.0
blinker==1.4
cached-property==1.5.1
celery==5.3.6
certifi==2019.11.28
cffi==1.15.1
channels==4.0.0
channels-auth-token-middlewares==1.1.0
channels-redis==4.1.0
chardet==3.0.4
click==8.1.3
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cryptography==42.0.5
cssselect==1.2.0
daphne==4.0.0
dj-database-url==2.1.0
Django==4.2
django-cors-headers==4.2.0
django-debug-toolbar==4.0.0
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
docker==4.1.0
dockerpty==0.4.1
docopt==0.6.2
drf-nested-routers==0.93.4
drf-yasg==1.21.7
entrypoints==0.3
exceptiongroup==1.1.3
filelock==3.11.0
httplib2==0.14.0
hyperlink==21.0.0
idna==2.8
importlib-metadata==6.6.0
incremental==22.10.0
inflection==0.5.1
itemadapter==0.8.0
itemloaders==1.0.6
itsdangerous==2.1.2
jmespath==1.0.1
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
kombu==5.3.5
lazr.uri==1.0.6
lxml==5.1.0
Markdown==3.4.3
MarkupSafe==2.1.2
more-itertools==4.2.0
msgpack==1.0.5
netifaces==0.10.4
packaging==24.0
parsel==1.9.0
pexpect==4.6.0
Pillow==9.5.0
prompt-toolkit==3.0.43
Protego==0.2.1
ptyprocess==0.7.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==24.1.0
pyrsistent==0.20.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2023.3
PyYAML==5.3.1
queuelib==1.6.2
redis==5.0.0
requests==2.22.0
requests-file==1.5.1
requests-unixsocket==0.2.0
service-identity==18.1.0
simplejson==3.16.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.4.4
swagger-spec-validator==3.0.3
twisted==24.3.0
txaio==23.1.1
typing-extensions==4.7.1
tzdata==2024.1
uritemplate==4.1.1
urllib3==1.25.8
vine==5.1.0
w3lib==2.1.1
wadllib==1.3.3
wcwidth==0.2.13
websocket-client==0.53.0
zipp==1.0.0
zope.interface==6.0
